+++ v1/backend/SPRING_LV_2_개노답_3계층/src/main/java/com/example/lent/controller/LentController.java
@@ -1,7 +1,8 @@
-package com.example.lent;
+package com.example.lent.controller;
 
 import com.example.lent.dto.LentRequest;
 import com.example.lent.dto.LentResponse;
+import com.example.lent.service.LentService;
 
 /*컨트롤러는 뭘까요?*/
 /*컨트롤러는 왜 인터페이스가 없어도 될까요?*/
