+++ v1/backend/SPRING_LV_2_개노답_3계층/src/main/java/com/example/lent/repository/LentHistoryRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.example.lent;
+package com.example.lent.repository;
 
 import com.example.lent.domain.LentHistory;
