+++ v1/backend/SPRING_LV_2_개노답_3계층/src/main/java/com/example/lent/service/LentService.java
@@ -1,4 +1,4 @@
-package com.example.lent;
+package com.example.lent.service;
 
 import com.example.lent.dto.LentRequest;
 import com.example.lent.dto.LentResponse;
